+++ snakeAI/ddpg.h
@@ -26,35 +26,35 @@ public:
                    int maxMemorySize = 4096,
                    int replaceTargetIter = 256,
                    int batchSize = 64);
-    void Perceive(std::vector<float>& state,
-                  float action,
-                  std::vector<float>& nextState,
-                  float reward,
+    void Perceive(std::vector<double>& state,
+                  double action,
+                  std::vector<double>& nextState,
+                  double reward,
                   bool done);
-    void SetSA(std::vector<float>& state, std::vector<float>& action);
+    void SetSA(std::vector<double>& state, std::vector<double>& action);
     void Forget();
-    int NoiseAction(std::vector<float>& state);
+    int NoiseAction(std::vector<double>& state);
     int RandomAction();
-    int GreedyAction(std::vector<float>& state);
-    int Action(std::vector<float>& state);
-    int MaxQ(std::vector<float>& q_value);
+    int GreedyAction(std::vector<double>& state);
+    int Action(std::vector<double>& state);
+    int MaxQ(std::vector<double>& q_value);
     void ExperienceReplay(Transition& x);
     void Learn(int optType = OPT_RMSPROP,
-               float actorLearningRate = 0.01f,
-               float criticLearningRate = 0.01f);
+               double actorLearningRate = 0.01,
+               double criticLearningRate = 0.01);
     void Save(const std::string& actorPara, const std::string& criticPara);
     void Load(const std::string& actorPara, const std::string& criticPara);
     int stateDim;
     int actionDim;
-    float gamma;
-    float alpha;
-    float beta;
+    double gamma;
+    double alpha;
+    double beta;
     int maxMemorySize;
     int batchSize;
-    float exploringRate;
+    double exploringRate;
     int learningSteps;
     int replaceTargetIter;
-    std::vector<float> sa;
+    std::vector<double> sa;
     BPNet actorMainNet;
     BPNet actorTargetNet;
     BPNet criticMainNet;
